requirements.txt
# Data
ccxt
python-binance
polars
pyarrow
pyyaml

# Utils
loguru
tqdm
python-dotenv
numpy
ipykernel
matplotlib
scikit-learn
statsmodels
seaborn
iisignature
optuna